move store to component mapping to its own file
