+++ src/components/Section/Identification/Identification.jsx
@@ -13,23 +13,7 @@ import {
   createPrintSubsectionViews
 } from '../generators'
 import navigation from './navigation'
-import ApplicantName from './ApplicantName'
-import ApplicantSSN from './ApplicantSSN'
-import ApplicantBirthPlace from './ApplicantBirthPlace'
-import ApplicantBirthDate from './ApplicantBirthDate'
-import OtherNames from './OtherNames'
-import Physical from './Physical'
-import ContactInformation from './ContactInformation'
-
-const storeToComponentMap = {
-  ApplicantBirthDate,
-  ApplicantBirthPlace,
-  ApplicantName,
-  ApplicantSSN,
-  Contacts: ContactInformation,
-  OtherNames,
-  Physical
-}
+import storeToComponentMap from './subsections'
 
 class Identification extends SectionElement {
   getSubsectionProps(subsection) {
